refactor: extract copy-mode state machine into dedicated module
